LayoutTests/imported/w3c/web-platform-tests/css/css-animations/style-animation-parsing-expected.txt


PASS Test animation name being empty.
PASS Test a non-conflicting animation name.
FAIL Test an animation name that is the same as a possible animation fill-mode. assert_equals: style1.animationName expected "none" but got "initial"
FAIL Test an animation name that is the same as a possible animation direction. assert_equals: style1.animationName expected "none" but got "initial"
FAIL Test an animation name that is the same as a possible running state. assert_equals: style1.animationName expected "none" but got "initial"
